Move algorithm modules to algo/
