+++ src/algo/simple_paths.rs
